Fixed broken links
Sorry for whatever reason, the links between individual pages were not working. I'm sure there was a reason this was coded like this, but I think we don't need it be linked in such an advance way. So did some basic linking just for submitting. Later we can revert back to the way it was.